refactor deskew tests
